Move networking_daemon to new class Endpoint
